fix: toggle theme functionality moved outside child component
