+++ src/components/header/index.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { ButtonIcon } from 'react-rainbow-components';
 import { MoonIcon, SunIcon, GitIcon } from '../icons';
@@ -6,38 +6,8 @@ import { Wrapper, LogoWrapper, Logo, TopActions, TopAction } from './styled';
 
 const gitUrl = 'https://github.com/nexxtway/rainbow-algolia-search/';
 
-const cyanDarkTheme = {
-    rainbow: {
-        palette: {
-            brand: '#44D7B6',
-            mainBackground: '#212121',
-        },
-    },
-};
-
-const lightTheme = {
-    rainbow: {
-        palette: {
-            brand: '#44D7B6',
-            mainBackground: '#f4f6f9',
-        },
-    },
-};
-
-const Header = ({ onSwitchTheme }) => {
-    const [icon, setIcon] = useState(<MoonIcon />);
-    const [currentTheme, setTheme] = useState('light');
-
-    const toogleTheme = () => {
-        if (currentTheme === 'dark') {
-            setIcon(<MoonIcon />);
-            setTheme('light');
-            return onSwitchTheme(lightTheme);
-        }
-        setIcon(<SunIcon />);
-        setTheme('dark');
-        return onSwitchTheme(cyanDarkTheme);
-    };
+const Header = ({ onSwitchTheme, theme }) => {
+    const icon = theme === 'light' ? <MoonIcon /> : <SunIcon />;
 
     return (
         <Wrapper>
@@ -46,7 +16,7 @@ const Header = ({ onSwitchTheme }) => {
             </LogoWrapper>
             <TopActions>
                 <TopAction href="#">
-                    <ButtonIcon icon={icon} size="large" onClick={toogleTheme} />
+                    <ButtonIcon icon={icon} size="large" onClick={onSwitchTheme} />
                 </TopAction>
                 <TopAction href={gitUrl} target="_blank">
                     <ButtonIcon icon={<GitIcon />} size="large" />
@@ -60,8 +30,10 @@ export default Header;
 
 Header.propTypes = {
     onSwitchTheme: PropTypes.func,
+    theme: PropTypes.string,
 };
 
 Header.defaultProps = {
     onSwitchTheme: () => {},
+    theme: '',
 };
