+++ src/components/layout/index.js
@@ -4,10 +4,10 @@ import { MainWrapper, Container } from './styled';
 import Header from '../header';
 import Footer from '../footer';
 
-const Layout = ({ children, onSwitchTheme }) => (
+const Layout = ({ children, onSwitchTheme, theme }) => (
     <>
         <MainWrapper>
-            <Header onSwitchTheme={onSwitchTheme} />
+            <Header onSwitchTheme={onSwitchTheme} theme={theme} />
             <Container>{children}</Container>
             <Footer />
         </MainWrapper>
@@ -19,9 +19,11 @@ export default Layout;
 Layout.propTypes = {
     children: PropTypes.node,
     onSwitchTheme: PropTypes.func,
+    theme: PropTypes.string,
 };
 
 Layout.defaultProps = {
     children: null,
     onSwitchTheme: () => {},
+    theme: '',
 };
